refactor(Sections): extract method of exploring
- Sectoins의 과중한 책임을 덜기 위해 Sections를 탐색하는 기능을 모두 외부 클래스에 위임
