+++ src/main/java/nextstep/subway/line/domain/sections/Sections.java
@@ -2,18 +2,13 @@ package nextstep.subway.line.domain.sections;
 
 import nextstep.subway.line.domain.exceptions.*;
 import nextstep.subway.line.domain.exceptions.InvalidSectionsActionException;
-import nextstep.subway.line.domain.exceptions.EndUpStationNotFoundException;
 import nextstep.subway.line.domain.exceptions.TargetSectionNotFoundException;
 import nextstep.subway.line.domain.exceptions.TooLongSectionException;
 
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-
-import static java.util.stream.Collectors.counting;
 
 @Embeddable
 public class Sections {
@@ -63,7 +58,7 @@ public class Sections {
 
     public boolean addNotEndSection(final Section newSection) {
         int originalSize = sections.size();
-        Section targetSection = this.findTargetSection(newSection);
+        Section targetSection = sectionExplorer.findTargetSection(this.sections, newSection);
         validateWhenAddNotEndSection(newSection, targetSection);
 
         OriginalSectionCalculator originalSectionCalculator = OriginalSectionCalculator.find(targetSection, newSection);
@@ -76,66 +71,24 @@ public class Sections {
         return (sections.size() == originalSize + 1);
     }
 
-    Section findEndUpSection() {
-        List<Long> singleStationIds = calculateSingleStationIds();
-
-        return this.sections.stream().filter(it -> it.isUpStationBelongsTo(singleStationIds))
-                .findFirst()
-                .orElseThrow(() -> new EndUpStationNotFoundException("상행종점역 구간을 찾을 수 없습니다."));
+    public boolean isInEndSection(final Section targetSection) {
+        return sectionExplorer.isInEndSection(sections, targetSection);
     }
 
-    Section findEndDownSection() {
-        List<Long> singleStationIds = calculateSingleStationIds();
-
-        return this.sections.stream().filter(it -> it.isDownStationBelongsTo(singleStationIds))
-                .findFirst()
-                .orElseThrow(() -> new EndUpStationNotFoundException("하행종점역 구간을 찾을 수 없습니다."));
-    }
-
-    Section findNextSection(final Section section) {
-        return this.sections.stream()
-                .filter(it -> it.isSameUpWithThatDown(section))
-                .findFirst()
-                .orElse(null);
-    }
-
-    Section findTargetSection(final Section newSection) {
-        Section targetSection = findSameWithUpStation(newSection);
-        if (targetSection == null) {
-            targetSection = findSameWithDownStation(newSection);
-        }
-        if (targetSection == null) {
-            throw new TargetSectionNotFoundException("새로운 구간을 추가할 수 있는 구간이 없습니다.");
-        }
-
-        return targetSection;
-    }
-
-    private List<Long> calculateSingleStationIds() {
-        return this.getStationIds().stream()
-                .collect(Collectors.groupingBy(Function.identity(), counting()))
-                .entrySet().stream()
-                .filter(it -> it.getValue() == 1L)
-                .map(Map.Entry::getKey)
-                .collect(Collectors.toList());
+    public List<Long> getStationIdsOrderBySection() {
+        return sectionExplorer.getStationIdsOrderBySection(sections);
     }
 
     public boolean isThisEndStation(Long stationId) {
-        return this.findEndUpSection().getUpStationId().equals(stationId) ||
-                this.findEndDownSection().getDownStationId().equals(stationId);
-    }
-
-    List<Section> findRelatedSections(final Long targetStationId) {
-        return sections.stream()
-                .filter(it -> it.isHasThisStation(targetStationId))
-                .collect(Collectors.toList());
+        return sectionExplorer.findEndUpSection(this.sections).getUpStationId().equals(stationId) ||
+                sectionExplorer.findEndDownSection(this.sections).getDownStationId().equals(stationId);
     }
 
     public boolean mergeSectionsByStation(final Long stationId) {
         int validTargetSize = 2;
         int originalSize = this.sections.size();
 
-        List<Section> relatedSections = findRelatedSections(stationId);
+        List<Section> relatedSections = sectionExplorer.findRelatedSections(this.sections, stationId);
         if (relatedSections.size() != validTargetSize) {
             throw new MergeSectionFailException("병합할 수 없는 중간역입니다.");
         }
@@ -151,7 +104,7 @@ public class Sections {
         int validTargetSize = 1;
         int originalSize = this.sections.size();
 
-        List<Section> relatedSections = findRelatedSections(stationId);
+        List<Section> relatedSections = sectionExplorer.findRelatedSections(this.sections, stationId);
         if(relatedSections.size() != validTargetSize) {
             throw new InvalidStationDeleteTryException("종점이 아닌 역을 종점 삭제 기능으로 제거할 수 없습니다.");
         }
@@ -166,7 +119,7 @@ public class Sections {
 
     private void validateWhenAddEndSection(final Section newSection) {
         validateIsInit();
-        if (!sectionExplorer.isInEndSection(this, newSection)) {
+        if (!sectionExplorer.isInEndSection(sections, newSection)) {
             throw new InvalidSectionsActionException("종점이 아닌 구간으로 종점 구간 추가를 수행할 수 없습니다.");
         }
     }
@@ -193,29 +146,9 @@ public class Sections {
     private boolean isAllStationsIn(final Section newSection) {
         List<Long> stationIds = newSection.getStationIds();
 
-        return this.getStationIds().stream()
+        return sectionExplorer.getStationIds(this.sections).stream()
                 .distinct()
                 .collect(Collectors.toList())
                 .containsAll(stationIds);
-    }
-
-    private List<Long> getStationIds() {
-        return sections.stream()
-                .flatMap(it -> it.getStationIds().stream())
-                .collect(Collectors.toList());
-    }
-
-    private Section findSameWithUpStation(final Section section) {
-        return this.sections.stream()
-                .filter(it -> it.isSameUpStation(section) && !it.isSameDownStation(section))
-                .findFirst()
-                .orElse(null);
-    }
-
-    private Section findSameWithDownStation(final Section section) {
-        return this.sections.stream()
-                .filter(it -> !it.isSameUpStation(section) && it.isSameDownStation(section))
-                .findFirst()
-                .orElse(null);
     }
 }
